+++ plugins/transport-nio/src/main/java/org/elasticsearch/http/nio/ByteBufUtils.java
@@ -23,6 +23,7 @@ import io.netty.buffer.CompositeByteBuf;
 import io.netty.buffer.Unpooled;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.BytesRefIterator;
+import org.elasticsearch.common.bytes.AbstractBytesReference;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.stream.StreamInput;
 
@@ -72,7 +73,7 @@ class ByteBufUtils {
         return new ByteBufBytesReference(buffer, buffer.readableBytes());
     }
 
-    private static class ByteBufBytesReference extends BytesReference {
+    private static class ByteBufBytesReference extends AbstractBytesReference {
 
         private final ByteBuf buffer;
         private final int length;
