+++ server/src/main/java/org/elasticsearch/common/bytes/ByteBufferReference.java
@@ -25,13 +25,13 @@ import java.nio.ByteBuffer;
 import java.util.Objects;
 
 /**
- * This is a {@link BytesReference} backed by a {@link ByteBuffer}. The byte buffer can either be a heap or
+ * This is a {@link AbstractBytesReference} backed by a {@link ByteBuffer}. The byte buffer can either be a heap or
  * direct byte buffer. The reference is composed of the space between the {@link ByteBuffer#position()} and
  * {@link ByteBuffer#limit()} at construction time. If the position or limit of the underlying byte buffer is
  * changed, those changes will not be reflected in this reference. Any changes to the underlying data in the
  * byte buffer will be reflected in this reference.
  */
-public class ByteBufferReference extends BytesReference {
+public class ByteBufferReference extends AbstractBytesReference {
 
     private final ByteBuffer buffer;
     private final int length;
