+++ server/src/main/java/org/elasticsearch/common/bytes/CompositeBytesReference.java
@@ -29,12 +29,12 @@ import java.util.Arrays;
 import java.util.Objects;
 
 /**
- * A composite {@link BytesReference} that allows joining multiple bytes references
+ * A composite {@link AbstractBytesReference} that allows joining multiple bytes references
  * into one without copying.
  *
  * Note, {@link #toBytesRef()} will materialize all pages in this BytesReference.
  */
-public final class CompositeBytesReference extends BytesReference {
+public final class CompositeBytesReference extends AbstractBytesReference {
 
     private final BytesReference[] references;
     private final int[] offsets;
