+++ server/src/test/java/org/elasticsearch/indices/IndicesRequestCacheTests.java
@@ -32,6 +32,7 @@ import org.apache.lucene.search.TermQuery;
 import org.apache.lucene.search.TopDocs;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.common.bytes.AbstractBytesReference;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
 import org.elasticsearch.common.lucene.index.ElasticsearchDirectoryReader;
@@ -455,7 +456,7 @@ public class IndicesRequestCacheTests extends ESTestCase {
         assertNotEquals(key1, key5);
     }
 
-    private class TestBytesReference extends BytesReference {
+    private class TestBytesReference extends AbstractBytesReference {
 
         int dummyValue;
         TestBytesReference(int dummyValue) {
