+++ test/framework/src/test/java/org/elasticsearch/test/AbstractXContentTestCaseTests.java
